fix: rename broken import
