+++ src/utils/index.js
@@ -1,5 +1,5 @@
 export {isLastIndex} from "./is-last-index";
-export {getGlobalWithKey} from "./globals";
+export {getGlobalWithKey, getGlobal} from "./globals";
 export {isMobile, isTablet, isDesktop} from "./breakpoint";
 export {validateSlug} from "./validate-slug";
 export {getMicrocopy, getSection} from "./content";
